Request-factory.class.ts rename
